Refactor TripReader and CitiBikeGraphService
