+++ graph-db-driver/src/main/java/se/kth/csc/moderndb/cbexplorer/domain/CitiBikeGraphService.java
@@ -4,15 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.neo4j.support.Neo4jTemplate;
 import org.springframework.transaction.annotation.Transactional;
 import se.kth.csc.moderndb.cbexplorer.parser.CitiBikeParser;
-import se.kth.csc.moderndb.cbexplorer.parser.CitiBikeReader;
-import se.kth.csc.moderndb.cbexplorer.parser.data.BikeData;
-import se.kth.csc.moderndb.cbexplorer.parser.data.StationData;
-import se.kth.csc.moderndb.cbexplorer.parser.data.TripData;
+import se.kth.csc.moderndb.cbexplorer.parser.TripReader;
 
 import java.io.File;
 import java.io.IOException;
 import java.text.ParseException;
-import java.util.Collection;
 
 /**
  * Created by mhotan on 4/13/14.
@@ -26,53 +22,26 @@ public class CitiBikeGraphService {
         new CitiBikeParser(new TripReader()).parse(file);
     }
 
-    /**
-     * Receives trip data from the parser and saves it to Neo4j.
-     */
-    public class TripReader implements CitiBikeReader {
-        @Override
-        public void addTrips(Collection<TripData> trips) {
-            for (TripData tripData : trips) {
-                Trip trip = new Trip(tripData.getStartTime(), tripData.getEndTime(), tripData.getUserType(),
-                        tripData.getUserBirthYear(), tripData.getUserGenderAsString());
-                Bike bike = obtainBike(tripData.getBikeData());
-                Station startStation = obtainStation(tripData.getStartStationData());
-                Station endStation = obtainStation(tripData.getEndStationData());
+    public void createTrip(Trip trip, Station startStation, Station endStation, Bike bike) {
+        trip.startedFrom = startStation;
+        trip.endedAt = endStation;
+        trip.bike = bike;
+        template.save(trip);
+    }
 
-                trip.startedFrom = startStation;
-                trip.endedAt = endStation;
-                trip.bike = bike;
-                template.save(trip);
-            }
-        }
+    public Bike findBike(long id) {
+        return template.findUniqueEntity(Bike.class, "id", id);
+    }
 
-        /**
-         * Find or create a Bike instance corresponding to the data.
-         * @param bikeData
-         * @return new or existing Bike with a matching id
-         */
-        private Bike obtainBike(BikeData bikeData) {
-            Bike bike = template.findUniqueEntity(Bike.class, "id", bikeData.getId());
-            if (bike == null) {
-                bike = new Bike(bikeData.getId());
-                template.save(bike);
-            }
-            return bike;
-        }
+    public void createBike(Bike bike) {
+        template.save(bike);
+    }
 
-        /**
-         * Find or create a Station instance corresponding to the data.
-         * @param stationData
-         * @return new or existing Station with a matching stationId
-         */
-        private Station obtainStation(StationData stationData) {
-            Station station = template.findUniqueEntity(Station.class, "stationId", stationData.getStationId());
-            if (station == null) {
-                station = new Station(stationData.getStationId(), stationData.getName(),
-                        stationData.getLongitude(), stationData.getLatitude());
-                template.save(station);
-            }
-            return station;
-        }
+    public Station findStation(long stationId) {
+        return template.findUniqueEntity(Station.class, "stationId", stationId);
+    }
+
+    public void createStation(Station station) {
+        template.save(station);
     }
 }
